Move around the span deferral code and storage
