+++ lib/rails/tracer.rb
@@ -1,4 +1,5 @@
 require "rails/span_helpers"
+require "rails/defer_notifications"
 require "rails/rack/tracer"
 require "rails/active_record/tracer"
 require "rails/active_support/cache/tracer"
@@ -22,7 +23,7 @@ module Rails
         ActionView::Tracer.instrument(tracer: tracer, active_span: active_span) if action_view
 
         # hold the requests until they can be written
-        @requests = {} if full_trace
+        Rails::Tracer::Defer.enable if full_trace
       end
 
       def disable
@@ -31,10 +32,6 @@ module Rails
         Rails::Rack::Tracer.disable
         ActionController::Tracer.disable
         ActionView::Tracer.disable
-      end
-
-      def requests
-        @requests
       end
     end
   end
